requirements.txt
torch~=2.5.1+cu124
numpy~=1.26.3
matplotlib~=3.10.0
pillow~=10.2.0
h5py~=3.12.1
nibabel~=5.3.2
scipy~=1.14.1
scikit-image~=0.25.0
medpy~=0.5.2
ptflops~=0.7.4
tqdm~=4.67.1
ml-collections~=1.0.0
torchsummary~=1.5.1
monai~=1.4.0
batchgenerators~=0.25.1
timm~=1.0.14
albumentations~=2.0.3
torchio~=0.20.4
opencv-python~=4.11.0.86
pywavelets~=1.8.0
argparse~=1.4.0
pandas~=2.2.3
tensorboardx~=2.6.2.2
loguru~=0.7.3
torchvision~=0.20.1+cu124